move BadgeComponent inside TimelineItem
