+++ stories/primer/timeline_item_component/badge_component_stories.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class Primer::TimelineItem::BadgeComponentStories < ViewComponent::Storybook::Stories
+class Primer::TimelineItemComponent::BadgeComponentStories < ViewComponent::Storybook::Stories
   layout "storybook_centered_preview"
 
   story(:with_items) do
