+++ test/components/component_test.rb
@@ -50,7 +50,6 @@ class PrimerComponentTest < Minitest::Test
     [Primer::TextComponent, {}],
     [Primer::TruncateComponent, {}],
     [Primer::TimelineItemComponent, {}, proc { |component| component.body { "Foo" } }],
-    [Primer::TimelineItem::BadgeComponent, { icon: "people" }],
     [Primer::TooltipComponent, { label: "More" }],
     [Primer::UnderlineNavComponent, {}]
   ].freeze
