+++ test/components/stories_test.rb
@@ -5,7 +5,8 @@ require "test_helper"
 class AllComponentsHaveStoriesTest < Minitest::Test
   EXCLUDED_COMPONENTS = [
     Primer::BaseComponent,
-    Primer::FlexItemComponent
+    Primer::FlexItemComponent,
+    Primer::TimelineItemComponent::BadgeComponent,
   ].freeze
 
   COMPONENTS = Primer::Component.descendants - EXCLUDED_COMPONENTS
